Reorganize design for the changelog in Align and Fine Tune view
