+++ app/assets/stylesheets/application.scss
@@ -123,6 +123,10 @@ a:hover {
   border-bottom: unset !important;
 }
 
+.borderless {
+  border-bottom: unset !important;
+}
+
 .mandatory-fields-notice{
   text-align: end;
 }
